Split challenge description into its own file
Often candidates will delete the README which makes it impossible to
know the exact challenge text they were working from. As we're likely to
iterate on this a bit in the coming months, this commit splits the
challenge description out from the README, with an instruction to
keep it around when submitting the code.